+++ spec/defines/mc/domain_spec.rb
@@ -5,7 +5,7 @@ describe 'sendmail::mc::domain' do
     let(:facts) { facts }
     let(:pre_condition) { 'include sendmail::service' }
 
-    context "on #{os} with domain foobar" do
+    context "on #{os} with domainname foobar" do
       let(:title) { 'foobar' }
 
       it {
@@ -18,7 +18,7 @@ describe 'sendmail::mc::domain' do
       }
     end
 
-    context "on #{os} with domain debian-mta" do
+    context "on #{os} with domainname debian-mta" do
       let(:title) { 'debian-mta' }
 
       it {
@@ -27,6 +27,22 @@ describe 'sendmail::mc::domain' do
         is_expected.to contain_concat__fragment('sendmail_mc-domain-debian-mta') \
           .with_content(%r{^DOMAIN\(`debian-mta'\)dnl$}) \
           .with_order('07') \
+          .that_notifies('Class[sendmail::makeall]')
+      }
+    end
+
+    context "on #{os} with domainname foobar as parameter" do
+      let(:title) { 'debian-mta' }
+      let(:params) do
+        { domainname: 'foobar' }
+      end
+
+      it {
+        is_expected.to contain_class('sendmail::makeall')
+
+        is_expected.to contain_concat__fragment('sendmail_mc-domain-foobar')
+          .with_content(%r{^DOMAIN\(`foobar'\)dnl$})
+          .with_order('07')
           .that_notifies('Class[sendmail::makeall]')
       }
     end